pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "lukim-gather-server"
version = "0.1.0"
description = "\"Lukim gather server\""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.11"
environs = {version = "^9.4.0", extras = ["django"]}
marshmallow = "^3.14.1"
psycopg2-binary = "^2.9.3"
uvicorn = {version = "^0.16.0", optional = true, extras = ["standard"]}
gunicorn = {version = "^20.1.0", optional = true}
graphene-django = "^2.15.0"
django-filter = "^21.1"
django-graphql-jwt = "^0.3.4"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
safety = "^1.10.3"
django-extensions = "^3.1.5"
Werkzeug = "^2.0.2"
pre-commit = "^2.16.0"
model-bakery = "^1.4.0"

[tool.poetry.extras]
asgi = ["uvicorn", "gunicorn"]

[tool.black]
extend-exclude = "migrations/"

[tool.isort]
skip_glob = ["__pycache__", "**/migrations"]
profile = "black"
